% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeDH.R
\name{ChangeDH}
\alias{ChangeDH}
\title{Changes in compound dry-hot events}
\usage{
ChangeDH(pre,tem,threp,thret)
}
\arguments{
\item{pre}{monthly precipitation of several decades}

\item{tem}{monthly temperature of several decades}

\item{threp}{Threshold of precipitation (e.g., 20th percentile)}

\item{thret}{Threshold of temperature (e.g., 80th percentile)}
}
\value{
The absolute/relative change for each month and annual occurrence
}
\description{
Assess absolute/relative changes in compound dry-hot (DH) events
at the monthly scale and event occurrences at the annual scale.
}
\examples{
pre=runif(1200, min = 0, max = 100) # 100-year monthly precipitation
tem=runif(1200, min = 0, max = 100) # 100-year monthly temperature
threp=20 #Specify the percentile-based threshold
thret=80 # Specify the percentile-based threshold
chg<-ChangeDH(pre,tem,threp,thret) #  chg$abschg,chg$relchg, chg$yrtotal
}
\references{
Hao, Z. et al. (2018). A multivariate approach for  statistical
           assessments of compound extremes. J. Hydrol., 565: 87-94.

Hao, Z. et al. (2019). A monitoring and prediction system
           for compound dry and hot events. Environ. Res. Lett., 14:114034.
}
