% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeWH.R
\name{ChangeWH}
\alias{ChangeWH}
\title{Changes in compound humid-hot events}
\usage{
ChangeWH(pre,tem,threp,thret)
}
\arguments{
\item{pre}{monthly precipitation of several decades}

\item{tem}{monthly temperature of several decades}

\item{threp}{Threshold of precipitation (e.g., 80th percentile)}

\item{thret}{Threshold of temperature (e.g., 80th percentile)}
}
\value{
The absolute/relative change for each month and annual occurrence
}
\description{
Assess changes in compound humid-hot (WH) occurrences
             at monthly and annual scales.
}
\examples{
pre=runif(1200, min = 0, max = 100) # 100-year monthly precipitation
tem=runif(1200, min = 0, max = 100) # 100-year monthly temperature
threp=80 #Specify the percentile-based threshold
thret=80 # Specify the percentile-based threshold
chg<-ChangeWH(pre,tem,threp,thret) #  chg$abschg,chg$relchg, chg$yrtotal
}
\references{
Hao, Z., AghaKouchak, A., Phillips, T.J., 2013. Changes
in concurrent monthly precipitation and temperature extremes.
Environ. Res. Lett. 8(3), 034014.
}
