% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfptng.R
\docType{data}
\name{dfptng}
\alias{dfptng}
\title{Monthly precipitation (pre), temperature (tem), Niño3.4 (nino) and GPP(gpp)}
\format{
A data frame with 792 rows (12 months × 66 years) and 4 variables:
\describe{
  \item{pre}{Monthly precipitation(unit:mm/month)}
  \item{tem}{Monthly average temperature (unit: °C)}
  \item{nino}{Monthly ENSO index (e.g., Nino3.4)}
  \item{gpp}{GPP}
}
}
\source{
The data were obtained from ERA-5, NOAA, and FLUXCOM.
}
\usage{
data(dfptng)
}
\description{
A dataset of monthly precipitation, temperature, and Niño3.4 index
 for 1951-2016 and monthly GPP for 1981–2016 (1951-1980 are NAN).
}
\keyword{datasets}
