% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DriverLogit.R
\name{DriverLogit}
\alias{DriverLogit}
\title{Assess potential driving factors of compound dry-hot events.}
\usage{
DriverLogit(Y,X)
}
\arguments{
\item{Y}{Occurrence of compound dry-hot events (0-1 binary variable)}

\item{X}{Climate index  as the driving factor of compound events (e.g., ENSO)}
}
\value{
slope parameter and associated p-value
}
\description{
Use logistic regression model to establish relationship between
climate indices (e.g., ENSO) and occurrences of compound dry-hot events.
}
\examples{
X=c(-0.7,-1.2,1.3,0.7,-0.6,1.1,-0.5,0.8,0.5,
-0.5,1.6,-1.8,-0.5,-1.4,-0.1,2.2,-0.7,-1.1, 0.6, -1.7) # climate indices
Y=c(0,0,1,1,0,0,0,0,0,0,1,0,1,0,0,1,0,0,0,0 )   # dry-hot occurrences
div<-DriverLogit(Y,X)  # div$slope and div$pvalue
}
\references{
Hao, Z. et al. (2019). A monitoring and prediction system for
            compound dry and hot events. Environ. Res. Lett., 14:114034.
}
