% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMF.R
\name{ELMF}
\alias{ELMF}
\title{Empirical likelihood multiplication factor (ELMF)}
\usage{
ELMF(mp,mt,threp,thret)
}
\arguments{
\item{mp}{Precipitation}

\item{mt}{Temperature}

\item{threp}{Threshold of precipitation (e.g., 50th percentile)}

\item{thret}{Threshold of temperature}
}
\value{
ELMF
}
\description{
Compute the joint probability based on the sample and the independent case
}
\examples{
mp=matrix(rnorm(120,0,1),ncol=1)
mt=matrix(rnorm(120,0,1),ncol=1)
threp=20
thret=80
res<-ELMF(mp,mt,threp,thret)

}
\references{
Zscheischler, J. and S. I. Seneviratne (2017). Dependence of drivers affects risks associated with compound events. Science Advances, 3(6): e1700263.
}
