% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Empdis1.R
\name{Empdis1}
\alias{Empdis1}
\title{Univariate empirical probability}
\usage{
Empdis1(mp)
}
\arguments{
\item{mp}{monthly precipitation}
}
\value{
The empirical probability
}
\description{
Compute univariate empirical probability
}
\examples{
mp=matrix(rnorm(120,0,1),ncol=1)
nd<-Empdis1(mp)
}
\references{
Hao, Z. et al., 2019a. Statistical prediction of the severity of compound dry-hot events based on El Ni??o-Southern Oscillation. J. Hydrol., 572, 243-250.
}
