% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDH.R
\name{GetDH}
\alias{GetDH}
\title{Occurrence of compound dry-hot events}
\usage{
GetDH(mp,mt,threp,thret)
}
\arguments{
\item{mp}{Precipitation}

\item{mt}{Temperature}

\item{threp}{Threshold of precipitation (e.g., 20th percentile)}

\item{thret}{Threshold of temperature (e.g., 80th percentile)}
}
\value{
The occurrence of compound dry-hot events (0-1 binary variable)
}
\description{
Get compound dry-hot (DH) events based on thresholds
precipitation and temperature, resulting in a 0-1The binary variable
}
\examples{
mp=matrix(rnorm(60,0,1),ncol=1) # prec. of a month (e.g., August,60 years)
mt=matrix(rnorm(60,0,1),ncol=1) # temp. of a  month (e.g., August,60 years)
threp=20 #Specify the percentile-based threshold
thret=80 #Specify the percentile-based threshold
DH<-GetDH(mp,mt,threp,thret) # 0-1 series
}
\references{
Hao, Z. et al. (2018). A multivariate approach for  statistical
assessments of compound extremes. J. Hydrol., 565: 87-94.

Hao, Z. et al. (2019). A monitoring and prediction system for
compound dry and hot events. Environ. Res. Lett., 14:114034.
}
