% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpactMGDH.R
\name{ImpactMGDH}
\alias{ImpactMGDH}
\title{Impacts of concurrent droughts and hot extremes}
\usage{
ImpactMGDH(mp,mt,mi,x0,y0,z0=NULL)
}
\arguments{
\item{mp}{monthly precipitation of a month (e.g., August)}

\item{mt}{monthly temperature of a month (e.g., August)}

\item{mi}{Impact variable of a month (e.g., GPP, Crop yield)}

\item{x0}{Initial condition of  SPI (e.g., -1.2, dry conditions)}

\item{y0}{Initial condition of  STI (e.g., 1.2, hot conditions)}

\item{z0}{Initial condition of standardized impact indicator(e.g.,0, low GPP)}
}
\value{
(1) Conditional mean/variance  at SPI=x0, STI=y0, here z0=NULL
(2)conditional probability of lower SII given lower SPI and higher STI
or P(SZI<=z0|SPI<=x0, STI>y0), here z0 should be specified
}
\description{
Use meta-Gaussian model to construct  conditional
distributions of the standardized impact variable (SII)  given SPI and STI
(represented by random variables of Z, X, and Y).
The output include (1) the conditional mean and variance of STI given SPI
and STI (e.g., X0=-1.2 and y0=1.2) (2)The conditional probability of
P(SII<=z0|SPI<=x0, STI>y0),  or lower GPP given dry-hot conditions
}
\examples{
mp=matrix(rnorm(60,0,1),ncol=1) # precipitation of a month (60 years)
mt=matrix(rnorm(60,0,1),ncol=1) # temperature of a month  (60 years)
mi=matrix(rnorm(60,0,1),ncol=1)   # impact of a month (60 years)
imp1<-ImpactMGDH(mp,mt,mi,-1.2,1.2,NULL) # imp1$mu, imp1$sig
imp2<-ImpactMGDH(mp,mt,mi,-1.2,1.2,0)# imp2$Condprob
}
\references{
Feng S.,Hao Z., et al. (2019). Probabilistic evaluation of the
          impact of compound dry-hot events on global maize yields.
          Sci. Total. Environ., 689: 1228-1234.

Hao Z., et al. (2018). A multivariate approach for statistical
assessments of compound extremes. J. Hydrol., 565: 87-94.
}
