% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionLGR.R
\name{PredictionLGR}
\alias{PredictionLGR}
\title{Prediction of occurrences of compound events}
\usage{
PredictionLGR(CI,Y01,CI0)
}
\arguments{
\item{CI}{Climate index (CI) as the driving factor of compound events (e.g., ENSO)}

\item{Y01}{Occurrences of compound dry-hot events (0-1 binvary variable) (L lead time)}

\item{CI0}{The antecedent climate indices (L lead time)}
}
\value{
Probability of occurrences
}
\description{
Fit the logistic regression model (LGR) based on occurrences of compound events (Y01) and climate indices (CI).The output is the predicted probability of compound event occurrence for the given climate index value CI0
}
\examples{
CI=c(-0.7,-1.2,1.3,0.7,-0.6,1.1,-0.5,0.8,0.5,-0.5,1.6,-1.8,-0.5,-1.4,-0.1,2.2,-0.7,-1.1, 0.6, -1.7)
Y01=c(0,0,1,1,0,0,0,0,0,0,1,0,1,0,0,1,0,0,0,0 )
PredictionLGR(CI,Y01,2)
}
\references{
Hao, Z. et al. (2019). Statistical prediction of the severity of compound dry-hot events based on ENSO . J. Hydrol., 572: 243-250.
}
