% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredLogit.R
\name{PredLogit}
\alias{PredLogit}
\title{Prediction of compound event occurrences}
\usage{
PredLogit(Y,X,type,x0=NULL)
}
\arguments{
\item{Y}{Occurrences of compound events (0-1 binary variable) (L lead time)}

\item{X}{Climate index (CI) as the driving factor (e.g., ENSO) with some lags}

\item{type}{type is the "performance" or "prediction"}

\item{x0}{Specified CI value based on which the prediction is issued}
}
\value{
(1) Probability of occurrences estimated at x0 if type="prediction",
shouLd input x0 (2) AUC values and predictions LOOCV if type="performance",
x0 is not needed
}
\description{
Fit the logistic regression model (LGR) based on occurrences of
compound events (Y) and climate index (CI).The output is predicted probability
of compound event occurrence for the given climate index value CI0 or the
prediction performance (using AUC) and predicted values based on LOOCV
}
\examples{
Y=as.matrix(c(0,0,1,1,0,0,0,0,0,0,1,0,1,0,0,1,0))
X=c(-0.7,-1.2,1.3,0.7,-0.6,1.1,-0.5,0.8,0.5,-0.5,
   1.6,-1.8,-0.5,-1.4,-0.1,2.2,-0.7) #  dry-hot (or other) 0-1 events
pred1<-PredLogit(Y,X,type="prediction",2) #pred1$predx0
pred2<-PredLogit(Y,X,type="performance") #pred2$aucvalue,pred2$predictions
}
\references{
Hao, Z. et al. (2019). Statistical prediction of the severity
      of compound dry-hot events based on ENSO . J. Hydrol., 572: 243-250.
}
