% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCEIDH.R
\name{SCEIDH}
\alias{SCEIDH}
\title{Standardized Compound Event Indicator (SCEI) for dry-hot events}
\usage{
SCEIDH(pre,tem,ts)
}
\arguments{
\item{pre}{monthly precipitation of several decades}

\item{tem}{monthly temperature of several decades}

\item{ts}{time scale}
}
\value{
The monthly SCEI series of several decades
}
\description{
Compute SCEI based on monthly  precipitation and temperature.
}
\examples{
pre=matrix(rnorm(120,0,1),ncol=1) # 10-year monthly precipitation
tem=matrix(rnorm(120,0,1),ncol=1) # 10-year monthly temperature
ts=1; # ts<=12 otherwise you should revise the function
SCEIDH(pre,tem,ts) #Generate monthly data
}
\references{
Hao, Z. et al., 2019. Statistical prediction of the severity of
            compound dry-hot events based on El Ni??o-Southern Oscillation.
             J. Hydrol., 572, 243-250.
}
