% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpactMG.R
\name{ImpactMG}
\alias{ImpactMG}
\title{Impact  under droughts and hot extremes}
\usage{
ImpactMG(P,T,Y)
}
\arguments{
\item{P}{Precipitation or drought indicator corresponding to the impact variable Y}

\item{T}{Temperature or heat indicator  corresponding to the impact variable Y}

\item{Y}{Impact variable (e.g., Crop yield)}
}
\description{
Use the meta-Gaussian model to construct the condtional distribuiton of
the impact variable (Y)  give drought and hot conditions P(Y|P,T)
}
\examples{
P=matrix(rnorm(60,0,1),ncol=1)
T=matrix(rnorm(60,0,1),ncol=1)
Y=matrix(rnorm(60,0,1),ncol=1)
ImpactMG(P,T,Y)

}
\references{
Feng, S.et al. (2019). Probabilistic evaluation of the impact of compound dry-hot events on global maize yields. Sci. Total. Environ., 689: 1228-1234.

Hao, Z.  et al. (2018). A multivariate approach for statistical assessments of compound extremes. J. Hydrol., 565: 87-94.
}
